extract data.api.tests.js from user.api.tests.js
